t031a5_middleware/requirements.txt
# Core dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
pydantic>=2.5.0
python-multipart>=0.0.6

# Audio processing
pyaudio>=0.2.11
speech-recognition>=3.10.0
pyttsx3>=2.90
numpy>=1.24.0
scipy>=1.11.0
librosa>=0.10.0

# Video processing
opencv-python>=4.8.0
pillow>=10.0.0
imageio>=2.31.0

# Motion control and robotics
unitree-legged-sdk>=3.8.0
transforms3d>=0.4.1

# Sensor data processing
scipy>=1.11.0
filterpy>=1.4.5
pyserial>=3.5

# LED control
rpi-ws281x>=5.0.0; platform_machine=="armv7l" or platform_machine=="aarch64"
neopixel>=6.3.0; platform_machine=="armv7l" or platform_machine=="aarch64"

# Database (optional)
sqlalchemy>=2.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.0

# Monitoring and metrics
prometheus-client>=0.19.0
psutil>=5.9.0

# Logging and configuration
coloredlogs>=15.0
pyyaml>=6.0
toml>=0.10.2

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
httpx>=0.25.0

# Development tools
black>=23.0.0
flake8>=6.0.0
mypy>=1.7.0
isort>=5.12.0
pre-commit>=3.5.0

# Utilities
click>=8.1.0
rich>=13.7.0
tqdm>=4.66.0
requests>=2.31.0
aiofiles>=23.2.0
watchdog>=3.0.0

# Security
cryptography>=41.0.0
pyjwt>=2.8.0
passlib[bcrypt]>=1.7.4

# Data serialization
orjson>=3.9.0
msgpack>=1.0.7

# Async utilities
aiohttp>=3.9.0
aioredis>=2.0.0

# System integration
setproctitle>=1.3.3
daemon>=1.2

# Optional ML/AI dependencies
# torch>=2.1.0
# torchvision>=0.16.0
# tensorflow>=2.14.0
# scikit-learn>=1.3.0
# opencv-contrib-python>=4.8